refactor(ApplicationTest.java): 클래스별 테스트 분리
ApplicationTest 클래스가 입출력에 대한 부분만 담당할 수 있게 분리했습니다
