+++ src/test/java/racingcar/MovingUnit.java
@@ -1,0 +1,6 @@
+package racingcar;
+
+public class MovingUnit {
+    public static final int MOVING_FORWARD = 4;
+    public static final int STOP = 3;
+}
